Connected API to React App
