+++ actcept/public/index.html
@@ -34,6 +34,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <p> Test </p>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
