+++ docker-compose.yml
@@ -1,15 +1,27 @@
 version: "3" 
 services:
-  server:
-    build: ./actcept
-    container_name: "actcept_container"
+  app:
+    container_name: react-app-prod
+    image: react-app-prod
+    build:
+      context: ./actcept
+      dockerfile: Dockerfile
+    ports: 
+      - 3000:3000
+    
+  api:
+    container_name: api_actcept
+    build: 
+      context: ./actcept-api
+      dockerfile: Dockerfile
     depends_on:
       - database
     ports:
-      - 3000:3000
+      - 3001:3001
     environment:
       - DB_HOSTNAME=myTestPostgres2
       - DB_PASSWORD=mysecretpassword
+
   database:
     image: postgres:latest
     container_name: myTestPostgres2
@@ -17,6 +29,5 @@ services:
       - POSTGRES_PASSWORD=mysecretpassword
     ports:
       - 5432:5432
-  
 
 # docker run --name my-postgres -p 5432:5432 -e POSTGRES_PASSWORD=mysecretpassword -d postgres
